Split program module into program/ package with mixins
