+++ a816/program/__init__.py
@@ -1,0 +1,7 @@
+"""`a816.program`: `Program` orchestrator + mixin handlers."""
+
+from __future__ import annotations
+
+from a816.program.core import Program
+
+__all__ = ["Program"]
